Implement random MC move selection (RandomTrigger)
Randomly select a single MC updater to be executed every time step from a set
of eligible ones. Updaters that correspond to randomly selectable MC moves
register themselves with the trigger, and ask it at the beginning of every
udpate() whether they have been selected for execution. Moves can optioanlly be
assigned a non-unit weight.

This C++ implementation serves as a prototype for a pure python implementation
going forward, but we need it as long as the main loop is executed from the C++
side (System class).

Minor change: change the semantics of System::addUpdater(). If there is an
Updater already with the same name, replace it. This change became necessary
since the HPMC integrator now creates an associated Updater (grid shift), and
we would like to reuse it if the integrator is replaced with a different one.